added float

+++ test101.py
@@ -14,7 +14,7 @@ while True:
     while True:
         input_1 = input("Input first number: ")
         try:
-            input_1 = int(input_1)
+            input_1 = float(input_1)
             break
         except:
             print("Invalid input! Ensure input is a number")
@@ -29,7 +29,7 @@ while True:
     while True:
         input_2 = input("Input second number: ")
         try:
-            input_2 = int(input_2)
+            input_2 = float(input_2)
             break
         except:
             print("Invalid input! Ensure input is a number")
